requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.0.0=pypi_0
albumentations=1.1.0=pypi_0
appnope=0.1.2=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.10.0=pypi_0
bleach=4.1.0=pypi_0
brotli=1.0.9=h3422bc3_6
brotli-bin=1.0.9=h3422bc3_6
bzip2=1.0.8=h3422bc3_4
ca-certificates=2021.10.8=h4653dfc_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py38h10201cd_2
cffi=1.15.0=py38hc67bbb8_0
charset-normalizer=2.0.12=pypi_0
click=8.1.0=pypi_0
coremltools=5.2.0=pypi_0
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.6.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.13=pypi_0
efficientnet-pytorch=0.6.3=pypi_0
entrypoints=0.4=pypi_0
executing=0.8.3=pypi_0
fonttools=4.31.2=py38h33210d7_0
freetype=2.10.4=h17b34a0_1
giflib=5.2.1=h27ca646_2
google-auth=2.6.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.44.0=pypi_0
humanize=4.0.0=pypi_0
idna=3.3=pypi_0
imageio=2.16.1=pypi_0
importlib-metadata=4.11.3=pypi_0
importlib-resources=5.6.0=pypi_0
imutils=0.5.4=pypi_0
ipykernel=6.10.0=pypi_0
ipython=8.2.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.0=pypi_0
jbig=2.1=h3422bc3_2003
jedi=0.18.1=pypi_0
jinja2=3.1.1=pypi_0
joblib=1.1.0=pypi_0
jpeg=9e=h3422bc3_0
jsonschema=4.4.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.2.0=pypi_0
jupyter-console=6.4.3=pypi_0
jupyter-core=4.9.2=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.1.0=pypi_0
kiwisolver=1.4.2=py38h0f71543_0
lcms2=2.12=had6a04f_0
lerc=3.0=hbdafb3b_0
libblas=3.9.0=13_osxarm64_openblas
libbrotlicommon=1.0.9=h3422bc3_6
libbrotlidec=1.0.9=h3422bc3_6
libbrotlienc=1.0.9=h3422bc3_6
libcblas=3.9.0=13_osxarm64_openblas
libcxx=13.0.1=h6a5c8ee_0
libdeflate=1.10=h3422bc3_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0.dev0=11_0_1_hf114ba7_23
libgfortran5=11.0.1.dev0=hf114ba7_23
liblapack=3.9.0=13_osxarm64_openblas
libopenblas=0.3.18=openmp_h5dd58f0_0
libpng=1.6.37=hf7e6567_2
libprotobuf=3.19.4=hccf11d3_0
libtiff=4.3.0=h77dc3b6_3
libwebp=1.2.2=h0d20362_0
libwebp-base=1.2.2=h3422bc3_1
libxcb=1.13=h9b22ae9_1004
libzlib=1.2.11=h90dfc92_1014
llvm-openmp=13.0.1=h455960f_1
lz4-c=1.9.3=hbdafb3b_1
markdown=3.3.6=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=py38h150bfb4_0
matplotlib-base=3.5.1=py38hb140015_0
matplotlib-inline=0.1.3=pypi_0
medpy=0.4.0=pypi_0
mistune=0.8.4=pypi_0
mpmath=1.2.1=pypi_0
munch=2.5.0=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.5.13=pypi_0
nbconvert=6.4.5=pypi_0
nbformat=5.2.0=pypi_0
ncurses=6.3=hc470f4d_0
nest-asyncio=1.5.4=pypi_0
networkx=2.7.1=pypi_0
nibabel=3.2.2=pypi_0
ninja=1.10.2=hc021e02_1
notebook=6.4.10=pypi_0
numpy=1.22.3=py38hf29d37f_0
oauthlib=3.2.0=pypi_0
opencv-python-headless=4.5.5.64=pypi_0
openjpeg=2.4.0=h062765e_1
openssl=3.0.2=h90dfc92_1
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.1=py38h3777fb4_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=py38h7ff1586_2
pip=22.0.4=pyhd8ed1ab_0
pretrainedmodels=0.7.4=pypi_0
prometheus-client=0.13.1=pypi_0
prompt-toolkit=3.0.28=pypi_0
protobuf=3.19.4=pypi_0
psutil=5.9.0=pypi_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pydicom=2.3.0=pypi_0
pygments=2.11.2=pypi_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyrsistent=0.18.1=pypi_0
python=3.8.13=hd3575e6_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.11.0=cpu_py38h1b6422d_0
pytorch-cpu=1.11.0=cpu_py38hd610c6a_0
pytz=2022.1=pyhd8ed1ab_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=pypi_0
qtconsole=5.3.0=pypi_0
qtpy=2.0.1=pypi_0
qudida=0.0.4=pypi_0
readline=8.1=hedafd6a_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-image=0.19.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
segmentation-models-pytorch=0.2.1=pypi_0
send2trash=1.8.0=pypi_0
setuptools=61.2.0=py38h10201cd_3
simpleitk=2.1.1=pypi_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h156473d_2
soupsieve=2.3.1=pypi_0
sqlite=3.37.1=h7e3ccbd_0
stack-data=0.2.0=pypi_0
sympy=1.10.1=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.13.3=pypi_0
testpath=0.6.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.3.25=pypi_0
timm=0.4.12=pypi_0
tk=8.6.12=he1e0b03_0
torchio=0.18.75=pypi_0
torchvision=0.12.0=pypi_0
tornado=6.1=py38hea4295b_2
tqdm=4.63.1=pypi_0
traitlets=5.1.1=pypi_0
typing_extensions=4.1.1=pyha770c72_0
unicodedata2=14.0.0=py38hea4295b_0
urllib3=1.26.9=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.1.0=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.6.0=pypi_0
wrapt=1.14.0=pypi_0
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.5=h642e427_1
zipp=3.7.0=pypi_0
zlib=1.2.11=h90dfc92_1014
zstd=1.5.2=h861e0a7_0
